Move SystemClock_Config to the same place the Olimex BSP has it.
