+++ hw/bsp/stm32f7discovery/src/hal_bsp.c
@@ -33,12 +33,8 @@
 #include <hal/hal_timer.h>
 
 #include <stm32f746xx.h>
-#include <stm32f7xx_hal_gpio_ex.h>
-#include <stm32f7xx_hal_flash.h>
-#include <stm32f7xx_hal_rcc.h>
-#include <stm32f7xx_hal_pwr.h>
-#include <stm32f7xx_hal_pwr_ex.h>
 #include <mcu/stm32f7_bsp.h>
+#include <stm32f7xx_hal_gpio_ex.h>
 
 #if MYNEWT_VAL(ETH_0)
 #include <stm32_eth/stm32_eth.h>
@@ -130,63 +126,12 @@ hal_bsp_core_dump(int *area_cnt)
     return dump_cfg;
 }
 
-/** System Clock Configuration.
- *
- * Configures CPU to run at 168 MHz.
- *
- * ((16 MHz / 10) * 210) / 2 = 168 MHz
- */
-void SystemClock_Config(void)
-{
-
-    RCC_OscInitTypeDef RCC_OscInitStruct;
-    RCC_ClkInitTypeDef RCC_ClkInitStruct;
-
-    /**Configure the main internal regulator output voltage
-    */
-    __HAL_RCC_PWR_CLK_ENABLE();
-
-    __HAL_PWR_VOLTAGESCALING_CONFIG(PWR_REGULATOR_VOLTAGE_SCALE2);
-
-    /* Initializes the CPU, AHB and APB busses clocks */
-    RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSI;
-    RCC_OscInitStruct.HSIState = RCC_HSI_ON;
-    RCC_OscInitStruct.HSICalibrationValue = 16;
-    RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-    RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSI;
-    RCC_OscInitStruct.PLL.PLLM = 10;
-    RCC_OscInitStruct.PLL.PLLN = 210;
-    RCC_OscInitStruct.PLL.PLLP = RCC_PLLP_DIV2;
-    RCC_OscInitStruct.PLL.PLLQ = 2;
-    if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK)
-    {
-        /* TODO: Throw error */
-    }
-
-    /* Initializes the CPU, AHB and APB busses clocks */
-    RCC_ClkInitStruct.ClockType = \
-        RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_PCLK1 | \
-        RCC_CLOCKTYPE_PCLK2;
-    RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-    RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-    RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV4;
-    RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV2;
-
-    if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_5) != HAL_OK)
-    {
-        /* TODO: Throw error */
-    }
-}
-
 void
 hal_bsp_init(void)
 {
     int rc;
 
     (void)rc;
-
-    SCB_EnableICache();
-    SystemClock_Config();
 
 #if MYNEWT_VAL(UART_0)
     rc = os_dev_create((struct os_dev *) &hal_uart0, "uart0",
